Altered a Admin service and Added a new admin service to cater both CEP and BAM
