--- components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.admin/src/main/java/org/wso2/carbon/analytics/dashboard/admin/data/Row.java
@@ -1,14 +1,0 @@
-package org.wso2.carbon.analytics.dashboard.admin.data;
-
-public class Row {
-    private Cell[] cells;
-
-    public Cell[] getCells() {
-        return cells;
-    }
-
-    public void setCells(Cell[] cells) {
-        this.cells = cells;
-    }
-
-}
