--- components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.admin/src/main/java/org/wso2/carbon/analytics/dashboard/admin/data/Table.java
@@ -1,22 +1,0 @@
-package org.wso2.carbon.analytics.dashboard.admin.data;
-
-public class Table {
-    private String name;
-    private Row[] rows;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Row[] getRows() {
-        return rows;
-    }
-
-    public void setRows(Row[] rows) {
-        this.rows = rows;
-    }
-}
