--- components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.admin/src/main/java/org/wso2/carbon/analytics/dashboard/admin/internal/ServiceHolder.java
@@ -1,16 +1,0 @@
-package org.wso2.carbon.analytics.dashboard.admin.internal;
-
-import org.wso2.carbon.analytics.dataservice.SecureAnalyticsDataService;
-
-public class ServiceHolder {
-
-    private static SecureAnalyticsDataService analyticsDataService;
-
-    public static void setAnalyticsDataService(SecureAnalyticsDataService analyticsDataService) {
-        ServiceHolder.analyticsDataService = analyticsDataService;
-    }
-
-    public static SecureAnalyticsDataService getAnalyticsDataService() {
-        return analyticsDataService;
-    }
-}
