+++ components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.batch.admin/pom.xml
@@ -1,36 +1,35 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
 <!--
- ~ Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- ~
- ~ Licensed under the Apache License, Version 2.0 (the "License");
- ~ you may not use this file except in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~      http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing, software
- ~ distributed under the License is distributed on an "AS IS" BASIS,
- ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- ~ See the License for the specific language governing permissions and
- ~ limitations under the License.
--->
+  ~ /**
+  ~  * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~  * Licensed under the Apache License, Version 2.0 (the "License");
+  ~  * you may not use this file except in compliance with the License.
+  ~  * You may obtain a copy of the License at
+  ~  *
+  ~  *      http://www.apache.org/licenses/LICENSE-2.0
+  ~  *
+  ~  * Unless required by applicable law or agreed to in writing, software
+  ~  * distributed under the License is distributed on an "AS IS" BASIS,
+  ~  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~  * See the License for the specific language governing permissions and
+  ~  * limitations under the License.
+  ~  */
+  -->
 
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
+        <artifactId>analytics-dashboard</artifactId>
         <groupId>org.wso2.carbon.analytics-common</groupId>
-        <artifactId>carbon-analytics-common</artifactId>
         <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../../pom.xml</relativePath>
     </parent>
-
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.analytics.dashboard.admin</artifactId>
+    <artifactId>org.wso2.carbon.analytics.dashboard.batch.admin</artifactId>
     <packaging>bundle</packaging>
-    <name>WSO2 Carbon - Analytics Dashboard Admin</name>
-    <description>Analytics Dashboard Admin</description>
-
+    <name>WSO2 Carbon - Batch Analytics Dashboard Admin</name>
+    <description>Batch Analytics Dashboard Admin</description>
 
     <dependencies>
         <!-- carbon -->
@@ -42,12 +41,22 @@
             <groupId>org.wso2.carbon</groupId>
             <artifactId>org.wso2.carbon.utils</artifactId>
         </dependency>
-        <!--<dependency>-->
-        <!--<groupId>com.fasterxml.jackson.jaxrs</groupId>-->
-        <!--<artifactId>jackson-jaxrs-json-provider</artifactId>-->
-        <!--<scope>runtime</scope>-->
-        <!--</dependency>-->
-
+        <!--carbon-analytics -->
+        <dependency>
+            <groupId>org.wso2.carbon.analytics</groupId>
+            <artifactId>org.wso2.carbon.analytics.dataservice</artifactId>
+            <version>1.0.3-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.analytics</groupId>
+            <artifactId>org.wso2.carbon.analytics.datasource.commons</artifactId>
+            <version>1.0.3-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.analytics</groupId>
+            <artifactId>org.wso2.carbon.analytics.dataservice.commons</artifactId>
+            <version>1.0.3-SNAPSHOT</version>
+        </dependency>
         <!-- Jackson -->
         <dependency>
             <groupId>org.codehaus.jackson</groupId>
@@ -64,13 +73,11 @@
             <artifactId>jackson-jaxrs</artifactId>
             <version>1.8.6</version>
         </dependency>
-
         <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
             <version>1.7.1</version>
         </dependency>
-
     </dependencies>
 
     <build>
@@ -89,7 +96,7 @@
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Carbon-Component>UIBundle</Carbon-Component>
                         <Export-Package>
-                            org.wso2.carbon.analytics.dashboard.admin.*,
+                            org.wso2.carbon.analytics.dashboard.batch.admin.*,
                         </Export-Package>
                         <Import-Package>
                             org.wso2.carbon.utils,
