+++ components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.batch.admin/src/main/java/org/wso2/carbon/analytics/dashboard/batch/admin/data/Table.java
@@ -1,0 +1,41 @@
+/*
+ * *
+ *  * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *      http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package org.wso2.carbon.analytics.dashboard.batch.admin.data;
+
+
+public class Table {
+    private String name;
+    private Row[] rows;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Row[] getRows() {
+        return rows;
+    }
+
+    public void setRows(Row[] rows) {
+        this.rows = rows;
+    }
+}
