+++ components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.batch.admin/src/main/java/org/wso2/carbon/analytics/dashboard/batch/admin/internal/ds/BatchAnalyticsDashboardAdminValueHolder.java
@@ -1,0 +1,34 @@
+/*
+ * *
+ *  * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *      http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package org.wso2.carbon.analytics.dashboard.batch.admin.internal.ds;
+
+import org.wso2.carbon.analytics.dataservice.SecureAnalyticsDataService;
+
+public class BatchAnalyticsDashboardAdminValueHolder {
+
+    private static SecureAnalyticsDataService analyticsDataService;
+
+    public static void setAnalyticsDataService(SecureAnalyticsDataService analyticsDataService) {
+        BatchAnalyticsDashboardAdminValueHolder.analyticsDataService = analyticsDataService;
+    }
+
+    public static SecureAnalyticsDataService getAnalyticsDataService() {
+        return analyticsDataService;
+    }
+}
