+++ components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.ui/pom.xml
@@ -42,6 +42,10 @@
             <artifactId>org.wso2.carbon.analytics.dashboard.stub</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.wso2.carbon.analytics-common</groupId>
+            <artifactId>org.wso2.carbon.analytics.dashboard.batch.stub</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
         </dependency>
