+++ components/analytics-dashboard/org.wso2.carbon.analytics.dashboard.ui/src/main/java/org/wso2/carbon/analytics/dashboard/ui/DashboardAdminClient.java
@@ -1,3 +1,17 @@
+/**
+ * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.wso2.carbon.analytics.dashboard.ui;
 
 import org.apache.axis2.AxisFault;
@@ -5,12 +19,19 @@ import org.apache.axis2.client.Options;
 import org.apache.axis2.client.ServiceClient;
 import org.apache.axis2.context.ConfigurationContext;
 import org.wso2.carbon.CarbonConstants;
+import org.wso2.carbon.analytics.dashboard.batch.stub.data.Cell;
+import org.wso2.carbon.analytics.dashboard.batch.stub.data.Row;
+import org.wso2.carbon.analytics.dashboard.batch.stub.data.Table;
 import org.wso2.carbon.analytics.dashboard.stub.DashboardAdminServiceStub;
-import org.wso2.carbon.analytics.dashboard.stub.data.*;
+import org.wso2.carbon.analytics.dashboard.stub.data.Column;
+import org.wso2.carbon.analytics.dashboard.stub.data.DataView;
+import org.wso2.carbon.analytics.dashboard.stub.data.Widget;
 import org.wso2.carbon.analytics.dashboard.ui.dto.ColumnDTO;
 import org.wso2.carbon.analytics.dashboard.ui.dto.DataViewDTO;
 import org.wso2.carbon.analytics.dashboard.ui.dto.TableDTO;
 import org.wso2.carbon.analytics.dashboard.ui.dto.WidgetDTO;
+
+import org.wso2.carbon.analytics.dashboard.batch.stub.BatchAnalyticsDashboardAdminServiceStub;
 import org.wso2.carbon.ui.CarbonUIUtil;
 
 import javax.servlet.ServletConfig;
@@ -29,12 +50,37 @@ public class DashboardAdminClient {
             ServletConfig config, HttpSession session,
             HttpServletRequest request)
             throws AxisFault {
+
         ConfigurationContext configContext = (ConfigurationContext) config.getServletContext()
                 .getAttribute(CarbonConstants.CONFIGURATION_CONTEXT);
         //Server URL which is defined in the server.xml
         String serverURL = CarbonUIUtil.getServerURL(config.getServletContext(),
                 session) + "DashboardAdminService.DashboardAdminServiceHttpsSoap12Endpoint";
         DashboardAdminServiceStub stub = new DashboardAdminServiceStub(configContext, serverURL);
+
+        String cookie = (String) session.getAttribute(org.wso2.carbon.utils.ServerConstants.ADMIN_SERVICE_COOKIE);
+
+        ServiceClient client = stub._getServiceClient();
+        Options option = client.getOptions();
+        option.setManageSession(true);
+        option.setProperty(org.apache.axis2.transport.http.HTTPConstants.COOKIE_STRING, cookie);
+
+        return stub;
+    }
+
+    public static BatchAnalyticsDashboardAdminServiceStub getDashboardBatchAnalyticsAdminService(
+            ServletConfig config, HttpSession session,
+            HttpServletRequest request)
+            throws AxisFault {
+
+        ConfigurationContext configContext = (ConfigurationContext) config.getServletContext()
+                .getAttribute(CarbonConstants.CONFIGURATION_CONTEXT);
+        //Server URL which is defined in the server.xml
+        String serverURL = CarbonUIUtil.getServerURL(config.getServletContext(),
+                session) + "BatchAnalyticsDashboardAdminService" +
+                ".BatchAnalyticsDashboardAdminServiceHttpsSoap12Endpoint";
+        BatchAnalyticsDashboardAdminServiceStub stub = new BatchAnalyticsDashboardAdminServiceStub
+                (configContext, serverURL);
 
         String cookie = (String) session.getAttribute(org.wso2.carbon.utils.ServerConstants.ADMIN_SERVICE_COOKIE);
 
@@ -75,7 +121,7 @@ public class DashboardAdminClient {
         DateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
         Date dateObj = null;
         try {
-//            dateObj = dateFormat.parse("23/09/2007");
+            //            dateObj = dateFormat.parse("23/09/2007");
             dateObj = dateFormat.parse(date);
         } catch (ParseException e) {
             e.printStackTrace();
