+++ pom.xml
@@ -294,6 +294,11 @@
                 <artifactId>org.wso2.carbon.analytics.dashboard.stub</artifactId>
                 <version>${carbon.analytics.common.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.wso2.carbon.analytics-common</groupId>
+                <artifactId>org.wso2.carbon.analytics.dashboard.batch.stub</artifactId>
+                <version>${carbon.analytics.common.version}</version>
+            </dependency>
 
             <!--carbon analytics common dependencies stop-->
