+++ service-stubs/org.wso2.carbon.analytics.dashboard.batch.stub/src/main/resources/BatchAnalyticsDashboardAdminService.wsdl
@@ -1,0 +1,105 @@
+<wsdl:definitions xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/" xmlns:ns1="http://org.apache.axis2/xsd" xmlns:ns="http://admin.batch.dashboard.analytics.carbon.wso2.org" xmlns:wsaw="http://www.w3.org/2006/05/addressing/wsdl" xmlns:http="http://schemas.xmlsoap.org/wsdl/http/" xmlns:ax23="http://data.admin.batch.dashboard.analytics.carbon.wso2.org/xsd" xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:mime="http://schemas.xmlsoap.org/wsdl/mime/" xmlns:soap12="http://schemas.xmlsoap.org/wsdl/soap12/" targetNamespace="http://admin.batch.dashboard.analytics.carbon.wso2.org">
+    <wsdl:documentation>BatchAnalyticsDashboardAdminService</wsdl:documentation>
+    <wsdl:types>
+        <xs:schema xmlns:ax24="http://data.admin.batch.dashboard.analytics.carbon.wso2.org/xsd" attributeFormDefault="qualified" elementFormDefault="qualified" targetNamespace="http://admin.batch.dashboard.analytics.carbon.wso2.org">
+            <xs:import namespace="http://data.admin.batch.dashboard.analytics.carbon.wso2.org/xsd"></xs:import>
+            <xs:element name="getRecords">
+                <xs:complexType>
+                    <xs:sequence>
+                        <xs:element minOccurs="0" name="tableName" nillable="true" type="xs:string"></xs:element>
+                        <xs:element minOccurs="0" name="timeFrom" type="xs:long"></xs:element>
+                        <xs:element minOccurs="0" name="timeTo" type="xs:long"></xs:element>
+                        <xs:element minOccurs="0" name="startIndex" type="xs:int"></xs:element>
+                        <xs:element minOccurs="0" name="recordCount" type="xs:int"></xs:element>
+                        <xs:element minOccurs="0" name="searchQuery" nillable="true" type="xs:string"></xs:element>
+                    </xs:sequence>
+                </xs:complexType>
+            </xs:element>
+            <xs:element name="getRecordsResponse">
+                <xs:complexType>
+                    <xs:sequence>
+                        <xs:element minOccurs="0" name="return" nillable="true" type="ax23:Table"></xs:element>
+                    </xs:sequence>
+                </xs:complexType>
+            </xs:element>
+        </xs:schema>
+        <xs:schema attributeFormDefault="qualified" elementFormDefault="qualified" targetNamespace="http://data.admin.batch.dashboard.analytics.carbon.wso2.org/xsd">
+            <xs:complexType name="Table">
+                <xs:sequence>
+                    <xs:element minOccurs="0" name="name" nillable="true" type="xs:string"></xs:element>
+                    <xs:element maxOccurs="unbounded" minOccurs="0" name="rows" nillable="true" type="ax23:Row"></xs:element>
+                </xs:sequence>
+            </xs:complexType>
+            <xs:complexType name="Row">
+                <xs:sequence>
+                    <xs:element maxOccurs="unbounded" minOccurs="0" name="cells" nillable="true" type="ax23:Cell"></xs:element>
+                </xs:sequence>
+            </xs:complexType>
+            <xs:complexType name="Cell">
+                <xs:sequence>
+                    <xs:element minOccurs="0" name="key" nillable="true" type="xs:string"></xs:element>
+                    <xs:element minOccurs="0" name="value" nillable="true" type="xs:string"></xs:element>
+                </xs:sequence>
+            </xs:complexType>
+        </xs:schema>
+    </wsdl:types>
+    <wsdl:message name="getRecordsRequest">
+        <wsdl:part name="parameters" element="ns:getRecords"></wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="getRecordsResponse">
+        <wsdl:part name="parameters" element="ns:getRecordsResponse"></wsdl:part>
+    </wsdl:message>
+    <wsdl:portType name="BatchAnalyticsDashboardAdminServicePortType">
+        <wsdl:operation name="getRecords">
+            <wsdl:input message="ns:getRecordsRequest" wsaw:Action="urn:getRecords"></wsdl:input>
+            <wsdl:output message="ns:getRecordsResponse" wsaw:Action="urn:getRecordsResponse"></wsdl:output>
+        </wsdl:operation>
+    </wsdl:portType>
+    <wsdl:binding name="BatchAnalyticsDashboardAdminServiceSoap11Binding" type="ns:BatchAnalyticsDashboardAdminServicePortType">
+        <soap:binding transport="http://schemas.xmlsoap.org/soap/http" style="document"></soap:binding>
+        <wsdl:operation name="getRecords">
+            <soap:operation soapAction="urn:getRecords" style="document"></soap:operation>
+            <wsdl:input>
+                <soap:body use="literal"></soap:body>
+            </wsdl:input>
+            <wsdl:output>
+                <soap:body use="literal"></soap:body>
+            </wsdl:output>
+        </wsdl:operation>
+    </wsdl:binding>
+    <wsdl:binding name="BatchAnalyticsDashboardAdminServiceSoap12Binding" type="ns:BatchAnalyticsDashboardAdminServicePortType">
+        <soap12:binding transport="http://schemas.xmlsoap.org/soap/http" style="document"></soap12:binding>
+        <wsdl:operation name="getRecords">
+            <soap12:operation soapAction="urn:getRecords" style="document"></soap12:operation>
+            <wsdl:input>
+                <soap12:body use="literal"></soap12:body>
+            </wsdl:input>
+            <wsdl:output>
+                <soap12:body use="literal"></soap12:body>
+            </wsdl:output>
+        </wsdl:operation>
+    </wsdl:binding>
+    <wsdl:binding name="BatchAnalyticsDashboardAdminServiceHttpBinding" type="ns:BatchAnalyticsDashboardAdminServicePortType">
+        <http:binding verb="POST"></http:binding>
+        <wsdl:operation name="getRecords">
+            <http:operation location="getRecords"></http:operation>
+            <wsdl:input>
+                <mime:content type="text/xml" part="parameters"></mime:content>
+            </wsdl:input>
+            <wsdl:output>
+                <mime:content type="text/xml" part="parameters"></mime:content>
+            </wsdl:output>
+        </wsdl:operation>
+    </wsdl:binding>
+    <wsdl:service name="BatchAnalyticsDashboardAdminService">
+        <wsdl:port name="BatchAnalyticsDashboardAdminServiceHttpsSoap11Endpoint" binding="ns:BatchAnalyticsDashboardAdminServiceSoap11Binding">
+            <soap:address location="https://192.168.2.101:9443/services/BatchAnalyticsDashboardAdminService.BatchAnalyticsDashboardAdminServiceHttpsSoap11Endpoint/"></soap:address>
+        </wsdl:port>
+        <wsdl:port name="BatchAnalyticsDashboardAdminServiceHttpsSoap12Endpoint" binding="ns:BatchAnalyticsDashboardAdminServiceSoap12Binding">
+            <soap12:address location="https://192.168.2.101:9443/services/BatchAnalyticsDashboardAdminService.BatchAnalyticsDashboardAdminServiceHttpsSoap12Endpoint/"></soap12:address>
+        </wsdl:port>
+        <wsdl:port name="BatchAnalyticsDashboardAdminServiceHttpsEndpoint" binding="ns:BatchAnalyticsDashboardAdminServiceHttpBinding">
+            <http:address location="https://192.168.2.101:9443/services/BatchAnalyticsDashboardAdminService.BatchAnalyticsDashboardAdminServiceHttpsEndpoint/"></http:address>
+        </wsdl:port>
+    </wsdl:service>
+</wsdl:definitions>
