+++ service-stubs/pom.xml
@@ -36,6 +36,7 @@
     <modules>
         <module>org.wso2.carbon.event.stream.stub</module>
         <module>org.wso2.carbon.analytics.dashboard.stub</module>
+        <module>org.wso2.carbon.analytics.dashboard.batch.stub</module>
     </modules>
 
 </project>
